gedit: v2.91.9
libgirepository is explicitly linked against, so gobject-introspection is a hard
dependency. The typlibs and girs are required for plugins.

libpeas python and gtk+ introspection is also required for most of the default
plugins. Since introspection is a hard dependency and it requires python itself
this shouldn't be a problem. The only additional dependency pulled in is
pygobject.