webkit[=1.6.3]: don't use --enable-debug-symbols
This is just a fancy way to add -g to C{,XX}FLAGS.

This change reduces the size of the resulting copy of webkit by 89%
(356.90MB->37.95MB), the size of the work dir by ~95% (7.4GB->354MB)
and the maximum memory consumption while building by ~42%
(2.53GB->1.48GB). All this data is taken with
build_options:symbols=compress, so the first number is off in favour of
the build with --enable-debug-symbols.